+++ packages/gatsby-plugin-mdx/src/__tests__/compile-mdx.ts
@@ -28,28 +28,28 @@ describe(`compiles MDX`, () => {
               "metadata": Object {},
               "processedMDX": "/*@jsxRuntime automatic @jsxImportSource react*/
             import {Fragment as _Fragment, jsx as _jsx, jsxs as _jsxs} from \\"react/jsx-runtime\\";
+            function _createMdxContent(props) {
+              const _components = Object.assign({
+                h1: \\"h1\\",
+                p: \\"p\\",
+                strong: \\"strong\\"
+              }, props.components), {Example} = _components;
+              if (!Example) _missingMdxReference(\\"Example\\", true);
+              return _jsxs(_Fragment, {
+                children: [_jsx(_components.h1, {
+                  children: \\"Hello\\"
+                }), \\"/n\\", _jsxs(_components.p, {
+                  children: [\\"This is \\", _jsx(_components.strong, {
+                    children: \\"MDX\\"
+                  })]
+                }), \\"/n\\", _jsx(Example, {})]
+              });
+            }
             function MDXContent(props = {}) {
               const {wrapper: MDXLayout} = props.components || ({});
               return MDXLayout ? _jsx(MDXLayout, Object.assign({}, props, {
-                children: _jsx(_createMdxContent, {})
-              })) : _createMdxContent();
-              function _createMdxContent() {
-                const _components = Object.assign({
-                  h1: \\"h1\\",
-                  p: \\"p\\",
-                  strong: \\"strong\\"
-                }, props.components), {Example} = _components;
-                if (!Example) _missingMdxReference(\\"Example\\", true);
-                return _jsxs(_Fragment, {
-                  children: [_jsx(_components.h1, {
-                    children: \\"Hello\\"
-                  }), \\"/n\\", _jsxs(_components.p, {
-                    children: [\\"This is \\", _jsx(_components.strong, {
-                      children: \\"MDX\\"
-                    })]
-                  }), \\"/n\\", _jsx(Example, {})]
-                });
-              }
+                children: _jsx(_createMdxContent, props)
+              })) : _createMdxContent(props);
             }
             export default MDXContent;
             function _missingMdxReference(id, component) {