+++ packages/gatsby-plugin-utils/src/polyfill-remote-file/utils/url-generator.ts
@@ -1,3 +1,4 @@
+import crypto from "crypto"
 import { basename, extname } from "path"
 import { URL } from "url"
 import { createContentDigest } from "gatsby-core-utils/create-content-digest"
@@ -9,8 +10,47 @@ const ORIGIN = `https://gatsbyjs.com`
 
 export enum ImageCDNUrlKeys {
   URL = `u`,
+  ENCRYPTED_URL = `eu`,
   ARGS = `a`,
   CONTENT_DIGEST = `cd`,
+}
+
+function encryptImageCdnUrl(
+  secretKey: string,
+  iv: string,
+  urlToEncrypt: string
+): string {
+  const randomPadding = crypto
+    .randomBytes(crypto.randomInt(32, 64))
+    .toString(`hex`)
+
+  const toEncrypt = `${randomPadding}:${urlToEncrypt}`
+  const cipher = crypto.createCipheriv(
+    `aes-256-ctr`,
+    Buffer.from(secretKey, `hex`),
+    Buffer.from(iv, `hex`)
+  )
+  const encrypted = cipher.update(toEncrypt)
+  const finalBuffer = Buffer.concat([encrypted, cipher.final()])
+
+  return finalBuffer.toString(`hex`)
+}
+
+function appendUrlParamToSearchParams(
+  searchParams: URLSearchParams,
+  url: string
+): void {
+  const key = process.env.IMAGE_CDN_ENCRYPTION_SECRET_KEY || ``
+  const iv = process.env.IMAGE_CDN_ENCRYPTION_IV || ``
+  const shouldEncrypt = !!(iv && key)
+
+  const paramName = shouldEncrypt
+    ? ImageCDNUrlKeys.ENCRYPTED_URL
+    : ImageCDNUrlKeys.URL
+
+  const finalUrl = shouldEncrypt ? encryptImageCdnUrl(key, iv, url) : url
+
+  searchParams.append(paramName, finalUrl)
 }
 
 export function generateFileUrl({
@@ -29,7 +69,7 @@ export function generateFileUrl({
     })}/${filenameWithoutExt}${fileExt}`
   )
 
-  parsedURL.searchParams.append(ImageCDNUrlKeys.URL, url)
+  appendUrlParamToSearchParams(parsedURL.searchParams, url)
 
   return `${parsedURL.pathname}${parsedURL.search}`
 }
@@ -52,7 +92,7 @@ export function generateImageUrl(
     )}/${filenameWithoutExt}.${imageArgs.format}`
   )
 
-  parsedURL.searchParams.append(ImageCDNUrlKeys.URL, source.url)
+  appendUrlParamToSearchParams(parsedURL.searchParams, source.url)
   parsedURL.searchParams.append(ImageCDNUrlKeys.ARGS, queryStr)
   parsedURL.searchParams.append(
     ImageCDNUrlKeys.CONTENT_DIGEST,