+++ packages/gatsby/src/utils/webpack/loaders/__tests__/fixtures/esm-default-function-anonymous.js
@@ -1,0 +1,3 @@
+"client export"
+
+export default function () {}