+++ packages/create-gatsby/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.24.0](https://github.com/gatsbyjs/gatsby/commits/create-gatsby@2.24.0/packages/create-gatsby) (2022-09-27)
+
+[🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v4.24)
+
+**Note:** Version bump only for package create-gatsby
+
+### [2.23.1](https://github.com/gatsbyjs/gatsby/commits/create-gatsby@2.23.1/packages/create-gatsby) (2022-09-22)
+
+**Note:** Version bump only for package create-gatsby
+
 ## [2.23.0](https://github.com/gatsbyjs/gatsby/commits/create-gatsby@2.23.0/packages/create-gatsby) (2022-09-13)
 
 [🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v4.23)