+++ packages/gatsby-plugin-mdx/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.3.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-plugin-mdx@4.3.0/packages/gatsby-plugin-mdx) (2022-09-27)
+
+[🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v4.24)
+
+#### Features
+
+- add partial hydration [#36485](https://github.com/gatsbyjs/gatsby/issues/36485) ([c073bbe](https://github.com/gatsbyjs/gatsby/commit/c073bbea40379adc62bd891da9f65ff2756b120b))
+
+#### Bug Fixes
+
+- Multiple root query error [#36525](https://github.com/gatsbyjs/gatsby/issues/36525) ([a922943](https://github.com/gatsbyjs/gatsby/commit/a9229439ec2de823148096991d5fab9b2f2bae6e))
+- Add React import when not defined [#36595](https://github.com/gatsbyjs/gatsby/issues/36595) ([34df27a](https://github.com/gatsbyjs/gatsby/commit/34df27a3bd278f49a9b81cdb59146a79e8841669))
+
+#### Chores
+
+- allow react/react-dom@experimental [#36533](https://github.com/gatsbyjs/gatsby/issues/36533) ([7ef4a3f](https://github.com/gatsbyjs/gatsby/commit/7ef4a3fe080d45e9edaff9f1d4deebd12a00ddbd))
+- Update "headings" instructions [#36593](https://github.com/gatsbyjs/gatsby/issues/36593) ([b1b83c1](https://github.com/gatsbyjs/gatsby/commit/b1b83c18a95a2ef79b6101b28d9295ef20b97d4b))
+
+### [4.2.1](https://github.com/gatsbyjs/gatsby/commits/gatsby-plugin-mdx@4.2.1/packages/gatsby-plugin-mdx) (2022-09-22)
+
+**Note:** Version bump only for package gatsby-plugin-mdx
+
 ## [4.2.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-plugin-mdx@4.2.0/packages/gatsby-plugin-mdx) (2022-09-13)
 
 [🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v4.23)